refactor: $utils/topic,bookへ整理 - sharedOrCreatedBy - connectOrCreateBook - connectOrCreateTopic
