+++ pages/book/import/index.tsx
@@ -8,25 +8,16 @@ import { useBooks } from "$utils/books";
 import type { BookSchema } from "$server/models/book";
 import type { SectionSchema } from "$server/models/book/section";
 import type { TopicSchema } from "$server/models/topic";
-import type { UserSchema } from "$server/models/user";
 import type { Query as BookEditQuery } from "../edit";
-import bookCreateBy from "$utils/bookCreateBy";
-import topicCreateBy from "$utils/topicCreateBy";
-import { createTopic } from "$utils/topic";
+import { connectOrCreateTopic } from "$utils/topic";
 import { pagesPath } from "$utils/$path";
 
 export type Query = BookEditQuery;
 
-const sharedOrCreatedBy = (author?: Pick<UserSchema, "id">) => (
-  book: BookSchema
-) => {
-  return book.shared || bookCreateBy(book, author);
-};
-
 function Import({ bookId, context }: Query) {
   const { data: session } = useSession();
   const book = useBook(bookId);
-  const books = useBooks();
+  const books = useBooks(session?.user);
   const router = useRouter();
   const bookEditQuery = { bookId, ...(context && { context }) };
   async function handleSubmit({
@@ -40,9 +31,8 @@ function Import({ bookId, context }: Query) {
 
     const connectOrCreateTopics = topics.map(async (topic) => {
       if (!session?.user) return Promise.reject();
-      if (topicCreateBy(topic, session.user)) return Promise.resolve(topic.id);
       // NOTE: 自身以外の作成したトピックに関しては影響を及ぼすのを避ける目的で複製
-      return createTopic(topic).then(({ id }) => id);
+      return connectOrCreateTopic(session.user, topic).then(({ id }) => id);
     });
     const ids = await Promise.all(connectOrCreateTopics);
     await updateBook({
@@ -83,12 +73,7 @@ function Import({ bookId, context }: Query) {
   if (!book) return <Placeholder />;
   if (!books) return <Placeholder />;
 
-  return (
-    <BookImport
-      books={books.filter(sharedOrCreatedBy(session?.user))}
-      {...handlers}
-    />
-  );
+  return <BookImport books={books} {...handlers} />;
 }
 
 function Router() {
