+++ utils/books.ts
@@ -1,7 +1,9 @@
 import useSWR from "swr";
 import type { BookSchema } from "$server/models/book";
+import type { UserSchema } from "$server/models/user";
 import { api } from "./api";
 import { revalidateBook } from "./book";
+import bookCreateBy from "./bookCreateBy";
 
 const key = "/api/v2/books";
 
@@ -12,7 +14,13 @@ async function fetchBooks(_: typeof key, page = 0): Promise<BookSchema[]> {
   return books;
 }
 
-export function useBooks() {
+const sharedOrCreatedBy = (author?: Pick<UserSchema, "id">) => (
+  book: BookSchema
+) => {
+  return book.shared || bookCreateBy(book, author);
+};
+
+export function useBooks(author: Pick<UserSchema, "id"> | undefined) {
   const { data } = useSWR<BookSchema[]>(key, fetchBooks);
-  return data;
+  return data?.filter(sharedOrCreatedBy(author));
 }
