+++ utils/topic.ts
@@ -1,6 +1,8 @@
 import useSWR, { mutate } from "swr";
 import type { TopicProps, TopicSchema } from "$server/models/topic";
+import type { UserSchema } from "$server/models/user";
 import { api } from "./api";
+import topicCreateBy from "./topicCreateBy";
 
 const key = "/api/v2/topic/{topic_id}";
 
@@ -20,6 +22,14 @@ export async function createTopic(body: TopicProps): Promise<TopicSchema> {
   })) as TopicSchema;
   await revalidateTopic(res.id, res);
   return res;
+}
+
+export async function connectOrCreateTopic(
+  user: Pick<UserSchema, "id">,
+  topic: TopicSchema
+) {
+  if (topicCreateBy(topic, user)) return topic;
+  return createTopic(topic);
 }
 
 export async function updateTopic({
