+++ utils/topics.ts
@@ -1,7 +1,9 @@
 import useSWR from "swr";
 import type { TopicSchema } from "$server/models/topic";
+import type { UserSchema } from "$server/models/user";
 import { api } from "./api";
 import { revalidateTopic } from "./topic";
+import topicCreateBy from "./topicCreateBy";
 
 const key = "/api/v2/topics";
 
@@ -12,7 +14,11 @@ async function fetchTopics(_: typeof key, page = 0): Promise<TopicSchema[]> {
   return topics;
 }
 
-export function useTopics() {
+const sharedOrCreatedBy = (creator?: Pick<UserSchema, "id">) => (
+  topic: TopicSchema
+) => topic.shared || topicCreateBy(topic, creator);
+
+export function useTopics(creator: Pick<UserSchema, "id"> | undefined) {
   const { data } = useSWR<TopicSchema[]>(key, fetchTopics);
-  return data;
+  return data?.filter(sharedOrCreatedBy(creator));
 }
